refactor: 이미지 업로드 폴더구조 수정
